Backend setup for BOM

+++ backend/OrderKraft/src/main/java/com/entity/BOM.java
@@ -17,7 +17,7 @@ public class BOM {
 	@Id
 	@GeneratedValue(strategy = GenerationType.SEQUENCE,generator = "bomseq")
 	@SequenceGenerator(name="bomseq",sequenceName = "bomseq",allocationSize = 1)
-	int bom_id;
+	Long bom_id;
 	String bom_name;
 	String remark;
 	
@@ -34,17 +34,17 @@ public class BOM {
 	public BOM() {
 		super();
 	}
-	public BOM(int bom_id, String bom_name, String remark, List<BOM_Material> materials) {
+	public BOM(Long bom_id, String bom_name, String remark, List<BOM_Material> materials) {
 		super();
 		this.bom_id = bom_id;
 		this.bom_name = bom_name;
 		this.remark = remark;
 		this.materials = materials;
 	}
-	public int getBom_id() {
+	public Long getBom_id() {
 		return bom_id;
 	}
-	public void setBom_id(int bom_id) {
+	public void setBom_id(Long bom_id) {
 		this.bom_id = bom_id;
 	}
 	public String getBom_name() {
